Fix project rename, re-work storage with synthetic project ids
Project data is stored in localStorage (a simple string -> string key
value store).

Previously we were encoding the app focus, project name, and sometimes
study area into the project key. This was tricky to reason about.

Now projects keys are opaque. A project has a ProjectID (uuid), and the
localStorage key path is a long keypath that includes that new
ProjectID.

Additional data like the project name and study area are stored in a
separate "index" object.

To be honest, it's a little more complicated than I'd like. I considered
just diving into indexdb, but that seemed like maybe *too* much (the api
is not synchronous).

Other notes:

We no longer add study_area_name to the savefile geojson. Since its in
the index, the app doesn't need it.

We do still need it for export though, so during export, studyAreaName
and projectName are added as foreign members in the web layer.
Note that previously we didn't explicitly add projectName - it was implicit in the filename.

TODO:

 -[ ] Migrate
 -[ ] Review
    -[x] update tests
    -[ ] minimize diff

DONE:

 -[x] Load from file
    -[x] Legacy
    -[x] Modern
 -[x] new from overpass (custom)
 -[x] CNT vs. LTN
 -[x] update loading text to use projectName
 -[x] load project from storage
 -[x] remove project
 -[x] create
 -[x] update
 -[x] save
 -[x] rename
 -[x] Download project file
    -[x] give better filename